requirements.txt
Django==3.0.3
Pillow==7.1.1
django-crispy-forms==1.9.0
django-summernote==0.8.11.6
django-taggit==1.2.0
gunicorn==20.0.4
whitenoise==5.0.1
django-import-export==2.0.2
djangorestframework==3.11.0
django-hitcount==1.3.1